+++ system/manager/process_config.py
@@ -134,6 +134,7 @@ elif TICI:
 procs += [
   PythonProcess("frogpilot_process", "frogpilot.frogpilot_process", always_run),
   NativeProcess("mapd", "frogpilot/navigation", ["./mapd"], always_run),
+  PythonProcess("offroad_wake_watcher", "frogpilot.system.offroad_wake_watcher", always_run, enabled=not PC),
   PythonProcess("apn_bridge", "tools.apn_bridge.apn_bridge", always_run),
   PythonProcess("speed_limit_filler", "frogpilot.system.speed_limit_filler", run_speed_limit_filler),
 ]